dev_requirements.txt
pytest==4.5.0
pytest-runner==5.1
python-rapidjson==0.7.1
urllib3-mock==0.3.3

-r requirements.txt
